Added modified account when doing a transfer.
If the transfer is to an account with other currency, normally the
amount is different because of the currency change. Now you can add a different
amount when doing the transfer.

It should be done in better way (checking the currency with ajax or
other way), but it's working...

There are known bugs when doing transfers :-/