[UPDATE] Upgraded to Rails 5.2.4.3
Modifications made in the sqlite database because of:
  https://api.rubyonrails.org/v5.2.3/classes/ActiveRecord/ConnectionAdapters/SQLite3Adapter.html#method-c-represent_boolean_as_integer

With "rails console" executed the next commands:
    Account.where("is_disabled = 't'").update_all(is_disabled: 1)
    Account.where("is_disabled = 'f'").update_all(is_disabled: 0)
    Movement.where("is_transfer = 't'").update_all(is_transfer: 1)
    Movement.where("is_transfer = 'f'").update_all(is_transfer: 0)

Those commands convert antique boolean format 't' into integer. Remember that I'm using PFM in development state with sqlite for my
personal use.

As mentioned in
  - https://github.com/rails/rails/issues/17062
  - https://github.com/rails/rails/issues/17062

this was made because of a dependency in a gem for sqlite3.